+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Current - Unreleased
 
+### Added
+ - Initial support for ROS2 message generation
+
+### Fixed
+ - The generated `char` type within rust is now u8.
+
 ## 0.5.2 - October 31, 2022
 
 ### Changed